+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.8.0-rc.31",
+  sdkVersion: "4.8.0-rc.32",
   genVersion: "2.881.2",
-  userAgent: "speakeasy-sdk/typescript 4.8.0-rc.31 2.881.2 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.8.0-rc.32 2.881.2 2.0 @orq-ai/node",
 } as const;