+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3675,7 +3675,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KPNFMKT1ZHQM74V91D8SS7BC"),
+  id: z.string().default("01KPNJFDN0T0Z11KC96HVZWERQ"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>