+++ packages/orq-rc/RELEASES.md
@@ -8389,3 +8389,13 @@ Based on:
 - [typescript v4.8.0-rc.31] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.31] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.31 - packages/orq-rc
+
+## 2026-04-20 13:53:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.8 (2.881.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.32] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.32] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.32 - packages/orq-rc